+++ module/handlebars-helpers.mjs
@@ -117,7 +117,7 @@ export function registerHandlebarsHelpers() {
     return item.getInventoryTitle();
   });
 
-  $$('toSacks', function (slotsInStones, displayUnit=true) {
+  $$('toSacks', function (slotsInStones, displayUnit = true) {
     const inSacks = convertToSacks(slotsInStones, SizeUnit.STONES);
 
     if (!displayUnit) return inSacks;
@@ -193,11 +193,31 @@ export function registerHandlebarsHelpers() {
     return set.has(value);
   });
 
+  $$('isDefeated', function(statuses = new Set()) {
+    return statuses.has('dead');
+  })
+
   $$('numberedFeature', function (feature) {
-    return ['replenish', 'flare', 'pocket', 'resistant'].includes(feature);
+    return ['area', 'replenish', 'flare', 'pocket', 'resistant'].includes(feature);
   });
 
   $$('replace', function (originalString = '', removedString = '', addedString = '') {
     return originalString.replace(removedString, addedString);
+  });
+
+  $$('featureDisplay', function (featureKey, parentSystem) {
+    let str = $l10n(`SDM.ItemFeature.${featureKey}Abbr`);
+    const numberedFeatures = ['area', 'replenish', 'flare', 'pocket', 'resistant'];
+    if (numberedFeatures.includes(featureKey)) {
+      const value = parentSystem[featureKey]?.value;
+      if (featureKey === 'area') {
+        const capitalized = value.charAt(0).toUpperCase() + value.slice(1);
+        const replacement = $l10n(`SDM.Area${capitalized}Abbr`);
+        str = str.replace('#', replacement);
+      } else {
+        str = str.replace('#', value);
+      }
+    }
+    return str;
   });
 }